api/requirements.txt
beautifulsoup4==4.9.3
boto3==1.16.49
celery==5.0.5
Django==3.1.5
django-rest-auth
django-allauth==0.44.0
django-cors-headers==3.6.0
django-mysql==3.10.0
django-storages==1.11.1
django-grappelli
djangorestframework==3.12.2
mysqlclient==2.0.3
redis